scholl88/Backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic-settings
mysql-connector-python
pymysql
pydantic-settings
python-multipart

